rename rule snooze badge component's file
