+++ x-pack/plugins/security_solution/public/detection_engine/components/rule_snooze_badge.tsx
